cleanup cache
